pyproject.toml
[tool.poetry]
name = "PikPakAPI"
version = "0.1.11"
description = "PikPakAPI"
authors = ["Quan666 <i@Rori.eMail>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/Quan666/PikPakAPI"
repository = "https://github.com/Quan666/PikPakAPI"
keywords = ["pikpak", "api"]

[tool.poetry.dependencies]
python = "^3.9.3"
httpx = ">=0.24.1"
oss2 = "^2.19.1"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
